mainloop: create common so POSIX and small OSes can share code
For timeout and idler processing we were having a lot of duplicated
code among POSIX and small OSes mainloops (e.g. RIOT). This patch moves
common code to a shared place and organizes it so all of the mainloop
implementations can use. It should also make it easier to add support
in the mainloops for handling addition/removal of event sources from
secondary threads.

Besides including the common mainloop header and compiling it together,
a mainloop implemenation using common needs to have the functions
described below.

   sol_mainloop_impl_lock()
   sol_mainloop_impl_unlock()
   sol_mainloop_impl_main_thread_check()
   sol_mainloop_impl_main_thread_notify()

     These functions need to either do nothing (in the single threaded
     case or no thread support) or implemented to allow the mainloop
     to take care of adding/removing event sources from other threads
     and processing them in the main thread.

     The lock and unlock ones will protect all mainloop data structures.
     The main thread check needs to return true if it's running in the
     context of the main thread, and false otherwise. The main thread
     notify function will be used to wake up the main thread whenever
     it's needed in common code. In the case there's no need to wake up
     the main thread this notify function should do nothing.

   sol_mainloop_impl_init()
   sol_mainloop_impl_shutdown()

     These must be implemented and must also call the respective init
     and shutdown functions from the common code.

   sol_mainloop_impl_iter()

     This function must be implemented to make one iteration of the
     specific mainloop. It needs to at least call into the timeout
     and idler process functions from common.

   sol_mainloop_impl_fd_add()
   sol_mainloop_impl_fd_del()
   sol_mainloop_impl_child_watch_add()
   sol_mainloop_impl_child_watch_del()

     These functions are not implemented in the common mainloop code.
     Thus the specific mainloop code needs to implement it, and in
     the case it's not supported alert the user with SOL_CRI() and
     return properly.
