+++ src/lib/common/Makefile
@@ -19,6 +19,7 @@ obj-core-$(MAINLOOP_GLIB)-extra-cflags += $(GLIB_CFLAGS)
 obj-core-$(MAINLOOP_GLIB)-extra-ldflags += $(GLIB_LDFLAGS)
 
 obj-core-$(MAINLOOP_POSIX) += \
+    sol-mainloop-common.o \
     sol-mainloop-impl-posix.o
 
 obj-core-$(MAINLOOP_RIOTOS) += \
